requirements.txt
streamlit
tensorflow
keras
google-cloud-bigquery
google-cloud-storage
google-auth
matplotlib